# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Plantit API
  version: 0.6.0
  description: |
    Contract-first definition for the Plantit backend read-path endpoints.
    Responses provide deterministic fixtures until persistence is available.
servers:
  - url: http://127.0.0.1:5581
paths:
...
  /api/villages:
    get:
      tags: [Villages]
      summary: List villages with optional filters applied.
      parameters:
        - in: query
          name: searchTerm
          schema:
            type: string
          description: Optional free-text search applied server side.
        - in: query
          name: climateZones
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          description: Climate zones to include.
        - in: query
          name: minHealth
          schema:
            type: number
            minimum: 0
            maximum: 1
          description: Minimum health score to surface.
      responses:
        '200':
          description: Villages with applied filters echoed back.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VillageListResponse'
    post:
      tags: [Villages]
      summary: Create a new village.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VillageWriteRequest'
      responses:
        '201':
          description: Created village payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VillageMutationResponse'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/villages/{villageId}:
    get:
      tags: [Villages]
      summary: Retrieve details for a specific village.
      parameters:
        - in: path
          name: villageId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Village detail payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VillageDetailResponse'
        '404':
          description: The requested village does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Villages]
      summary: Update an existing village.
      parameters:
        - in: path
          name: villageId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VillageUpdateRequest'
      responses:
        '200':
          description: Updated village payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VillageMutationResponse'
        '404':
          description: The requested village does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The village has been modified since the provided version token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Villages]
      summary: Delete a village using optimistic concurrency.
      parameters:
        - in: path
          name: villageId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VillageDeleteRequest'
      responses:
        '200':
          description: Deletion acknowledgement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VillageDeleteResponse'
        '404':
          description: The requested village does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The village has been modified since the provided version token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/villages/{villageId}/plants:
    get:
      tags: [Plants]
      summary: Retrieve plants within a given village.
      parameters:
        - in: path
          name: villageId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plant list scoped to the village.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VillagePlantListResponse'
        '404':
          description: The requested village does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/plants:
    post:
      tags: [Plants]
      summary: Create a plant within a village.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantCreateRequest'
      responses:
        '201':
          description: Created plant payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantMutationResponse'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The referenced village does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/plants/{plantId}:
    get:
      tags: [Plants]
      summary: Retrieve the plant modal payload for a single plant.
      parameters:
        - in: path
          name: plantId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plant detail with timeline events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantDetailResponse'
        '404':
          description: The requested plant does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Plants]
      summary: Update an existing plant.
      parameters:
        - in: path
          name: plantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantUpdateRequest'
      responses:
        '200':
          description: Updated plant payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantMutationResponse'
        '404':
          description: The requested plant does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The plant has been modified since the provided version token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Plants]
      summary: Delete a plant using optimistic concurrency.
      parameters:
        - in: path
          name: plantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantDeleteRequest'
      responses:
        '200':
          description: Deletion acknowledgement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantDeleteResponse'
        '404':
          description: The requested plant does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The plant has been modified since the provided version token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/import:
    post:
      tags: [Import/Export]
      summary: Submit an import preview for validation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportPreviewRequest'
      responses:
        '202':
          description: Import preview accepted for processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreviewResponse'
...
components:
  schemas:
...
    VillageListResponse:
      type: object
      required: [villages, appliedFilters]
      properties:
        villages:
          type: array
          items:
            $ref: '#/components/schemas/VillageSummary'
        appliedFilters:
          $ref: '#/components/schemas/VillageFilterState'
    VillageSummary:
      type: object
      required: [id, name, climate, plantCount, healthScore, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        climate:
          type: string
        plantCount:
          type: integer
          minimum: 0
        healthScore:
          type: number
          minimum: 0
          maximum: 1
        updatedAt:
          type: string
          format: date-time
    VillageFilterState:
      type: object
      required: [searchTerm, climateZones]
      properties:
        searchTerm:
          type: string
        climateZones:
          type: array
          items:
            type: string
        minHealth:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
    VillageDetailResponse:
      type: object
      required: [village]
      properties:
        village:
          $ref: '#/components/schemas/VillageDetail'
...
    VillagePlantListResponse:
      type: object
      required: [village, plants]
      properties:
        village:
          $ref: '#/components/schemas/VillageSummary'
        plants:
          type: array
          items:
            $ref: '#/components/schemas/PlantListItem'
...
    PlantDetailResponse:
      type: object
      required: [plant, timeline]
      properties:
        plant:
          $ref: '#/components/schemas/PlantDetail'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/PlantEvent'
    VillageWriteRequest:
      type: object
      required: [name, climate, healthScore]
      properties:
        name:
          type: string
        climate:
          type: string
        healthScore:
          type: number
          minimum: 0
          maximum: 1
        description:
          type: string
          nullable: true
        establishedAt:
          type: string
          format: date
          nullable: true
        irrigationType:
          type: string
          nullable: true
    VillageUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/VillageWriteRequest'
        - type: object
          required: [updatedAt]
          properties:
            updatedAt:
              type: string
              format: date-time
    VillageDeleteRequest:
      type: object
      required: [updatedAt]
      properties:
        updatedAt:
          type: string
          format: date-time
    VillageMutationResponse:
      type: object
      required: [village]
      properties:
        village:
          $ref: '#/components/schemas/VillageDetail'
    VillageDeleteResponse:
      type: object
      required: [status, villageId, updatedAt]
      properties:
        status:
          type: string
          enum: [deleted]
        villageId:
          type: string
        updatedAt:
          type: string
          format: date-time
...
    PlantCreateRequest:
      allOf:
        - $ref: '#/components/schemas/PlantWriteRequest'
        - type: object
          required: [villageId]
          properties:
            villageId:
              type: string
    PlantUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/PlantWriteRequest'
        - type: object
          required: [updatedAt]
          properties:
            updatedAt:
              type: string
              format: date-time
    PlantDeleteRequest:
      type: object
      required: [updatedAt]
      properties:
        updatedAt:
          type: string
          format: date-time
    PlantMutationResponse:
      type: object
      required: [plant]
      properties:
        plant:
          $ref: '#/components/schemas/PlantDetail'
        village:
          $ref: '#/components/schemas/VillageSummary'
          nullable: true
    PlantDeleteResponse:
      type: object
      required: [status, plantId, updatedAt]
      properties:
        status:
          type: string
          enum: [deleted]
        plantId:
          type: string
        updatedAt:
          type: string
          format: date-time
        village:
          $ref: '#/components/schemas/VillageSummary'
          nullable: true
...
    ImportPreviewRequest:
      type: object
      required: [schemaVersion, summary]
      properties:
        schemaVersion:
          type: integer
          minimum: 1
        summary:
          type: object
          description: Free-form summary counts generated by the client.
          additionalProperties: true
    ImportPreviewResponse:
      type: object
      required: [status, schemaVersion, summary, message]
      properties:
        status:
          type: string
          enum: [accepted]
        schemaVersion:
          type: integer
          minimum: 1
        summary:
          type: object
          additionalProperties: true
        message:
          type: string
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string